+++ go.mod
@@ -1,3 +1,1 @@
-module git@github.com:Asmmund/go_blockchain.git
-
-require rsc.io/quote v1.5.2 // indirect
+module github.com/Asmmund/go_blockchain
